Revert "feat: 포스트 태그 추가"
This reverts commit efe0b122ba6e60c61ba7194b837b49e891216025.
