+++ src/app/books/page.tsx
@@ -4,7 +4,6 @@ import { PostListSummary } from '@/components/PostListSummary';
 import { POST_LIST_SUMMARY } from '@/constants';
 import { getAllPosts } from '@/libs/api';
 import { Metadata } from 'next';
-import { PostTagTabs } from '@/components/PostTagTabs';
 import { openGraph, twitter } from '../shared-metadata';
 
 export const metadata: Metadata = {
@@ -29,7 +28,6 @@ export default function Books() {
     <Template.Root>
       <Template.Main>
         <PostListSummary type="books" posts={posts} />
-        <PostTagTabs posts={posts} />
         <PostList type="books" posts={posts} />
       </Template.Main>
     </Template.Root>
