--- src/components/PostListItem/index.ts
@@ -1,1 +1,0 @@
-export { default as PostListItem } from './PostListItem';
