--- src/components/PostTag/index.ts
@@ -1,1 +1,0 @@
-export { default as PostTag } from './PostTag';
