--- src/components/PostTagTabs/index.ts
@@ -1,1 +1,0 @@
-export { default as PostTagTabs } from './PostTagTabs';
